+++ production_changes/pending/changeset23.sql
@@ -37,7 +37,7 @@ BEGIN
 	AND ser.lodging_datetime BETWEEN to_date('''|| $1 ||''',''yyyy-mm-dd'')  and to_date('''|| $2 ||''',''yyyy-mm-dd'')
 	GROUP BY nt.request_type)
 	--standard deviation is a measure of the average distance of the values in the data set from their mean
-	SELECT request_type, service_count, total_time, round(total_time/service_count::float) AS average_time, frequent_day, round(stddev) as stddev, minimum, maximum,
+	SELECT request_type, service_count, total_time, round(total_time/service_count::float)::float AS average_time, frequent_day, round(stddev) as stddev, minimum, maximum,
 	 (maximum - minimum) as rng
 	FROM results_table order by request_type';
